refactor(api): extract power shelf controller crate
Move the power shelf state controller implementation out of the API crate into
a new =carbide-power-shelf-controller= crate.

Update API setup and power shelf controller tests to use the extracted crate
while preserving existing behavior. This continues separating controller code
from the API crate.
