+++ crates/api/src/state_controller/mod.rs
@@ -18,6 +18,5 @@
 pub mod common_services;
 pub(crate) mod external_service_error;
 pub mod machine;
-pub mod power_shelf;
 
 pub use ::state_controller::{controller, db_write_batch, state_change_emitter, state_handler};
