+++ crates/api/src/tests/power_shelf_state_controller/mod.rs
@@ -18,6 +18,9 @@
 use std::sync::Arc;
 use std::time::Duration;
 
+use carbide_power_shelf_controller::context::PowerShelfStateHandlerServices;
+use carbide_power_shelf_controller::handler::PowerShelfStateHandler;
+use carbide_power_shelf_controller::io::PowerShelfStateControllerIO;
 use db::power_shelf as db_power_shelf;
 use model::power_shelf::PowerShelfControllerState;
 use rpc::forge::forge_server::Forge;
@@ -25,9 +28,6 @@ use state_controller::config::IterationConfig;
 use tokio_util::sync::CancellationToken;
 
 use crate::state_controller::controller::StateController;
-use crate::state_controller::power_shelf::context::PowerShelfStateHandlerServices;
-use crate::state_controller::power_shelf::handler::PowerShelfStateHandler;
-use crate::state_controller::power_shelf::io::PowerShelfStateControllerIO;
 use crate::tests::common;
 use crate::tests::common::api_fixtures::create_test_env;
 mod error_state;
