+++ crates/power-shelf-controller/src/error_state.rs
@@ -23,7 +23,7 @@ use state_controller::state_handler::{
     StateHandlerContext, StateHandlerError, StateHandlerOutcome,
 };
 
-use crate::state_controller::power_shelf::context::PowerShelfStateHandlerContextObjects;
+use crate::context::PowerShelfStateHandlerContextObjects;
 
 /// Handles the Error state for a power shelf.
 ///
