+++ crates/power-shelf-controller/src/metrics.rs
